+++ test/autogenerated_all/models/todos/generated.rs
@@ -10,7 +10,7 @@ type ConnectionType = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionMan
 
 #[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, AsChangeset, Selectable)]
 #[diesel(table_name=todos, primary_key(id))]
-pub struct Todo {
+pub struct Todos {
     pub id: i32,
     pub created_at: chrono::NaiveDateTime,
 }
@@ -18,7 +18,7 @@ pub struct Todo {
 
 #[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, AsChangeset, Default)]
 #[diesel(table_name=todos)]
-pub struct UpdateTodo {
+pub struct UpdateTodos {
     pub created_at: Option<chrono::NaiveDateTime>,
 }
 
@@ -33,7 +33,7 @@ pub struct PaginationResult<T> {
     pub num_pages: i64,
 }
 
-impl Todo {
+impl Todos {
 
     pub fn create(db: &mut ConnectionType) -> QueryResult<Self> {
         use crate::schema::todos::dsl::*;
@@ -65,7 +65,7 @@ impl Todo {
         })
     }
 
-    pub fn update(db: &mut ConnectionType, param_id: i32, item: &UpdateTodo) -> QueryResult<Self> {
+    pub fn update(db: &mut ConnectionType, param_id: i32, item: &UpdateTodos) -> QueryResult<Self> {
         use crate::schema::todos::dsl::*;
 
         diesel::update(todos.filter(id.eq(param_id))).set(item).get_result(db)